Add utility function for merging class values
This commit adds a new utility function called `cn` to the `utils.ts` file. The `cn` function takes in multiple class values and uses the `clsx` and `tailwind-merge` libraries to merge them into a single class string. This utility function will be useful for dynamically applying CSS classes in the application.